update the client and server req-res on Auth App setup
