+++ mobile/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon ./src/index.ts",
-    "build": "tsc -p .",
+    "build": "rimraf lib && tsc -p .",
     "start": "node ./lib/index.js"
   },
   "keywords": [],
