+++ mobile/src/controller/auth-app-controller.ts
@@ -3,7 +3,6 @@ import { validationResult } from 'express-validator';
 
 export const authAppSetupController = async (req: Request, res: Response) => {
   const error = validationResult(req);
-
   if (!error.isEmpty()) {
     const errorMessage = error.array()[0].msg;
     res.status(400).json({ message: errorMessage });
@@ -11,6 +10,8 @@ export const authAppSetupController = async (req: Request, res: Response) => {
   }
 
   const { address } = req.body;
+  const { notifiToken } = req.body;
 
-  res.status(200).json({ address: address });
+  //TO-DO: store the address to Chain
+  res.status(200).json({ address: address, notifiToken: notifiToken});
 };
