--- mobile/src/router/auth-router.js
@@ -1,42 +1,0 @@
-"use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || function (mod) {
-    if (mod && mod.__esModule) return mod;
-    var result = {};
-    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
-    __setModuleDefault(result, mod);
-    return result;
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const express_1 = require("express");
-const authAppController = __importStar(require("../controller/auth-app-controller"));
-const express_validator_1 = require("express-validator");
-const router = (0, express_1.Router)();
-router.put('/setup', [
-    (0, express_validator_1.body)('name').trim().isLength({ min: 3 }).withMessage('Too small name'),
-    (0, express_validator_1.body)('email')
-        .trim()
-        .isEmail()
-        .withMessage('Please enter a valid email')
-        .normalizeEmail(),
-    (0, express_validator_1.body)('password')
-        .trim()
-        .isLength({ min: 5 })
-        .withMessage('Too small password'),
-], authAppController.authAppSetupController);
-exports.default = router;
