+++ mobile/src/router/auth-router.ts
@@ -4,11 +4,8 @@ import { body } from 'express-validator';
 
 const router = Router();
 
-router.put(
+router.post(
   '/setup',
-  [
-    body('address').trim().isLength({ min: 3 }).withMessage('Not valid address'),
-  ],
   authAppController.authAppSetupController
 );
